created test classes for each business-case
